fuse/runtest: make removal of webport file soft
previously the runtests suite removed the webport file created by
tahoe create-client in all but the first node.  now that the node config
is in tahoe.cfg by default this file might not exist.

+++ contrib/fuse/runtests.py
@@ -291,7 +291,8 @@ class SystemTest (object):
             self.weburl = "http://127.0.0.1:%d/" % (self.port,)
             print self.weburl
         else:
-            os.remove(webportpath)
+            if os.path.exists(webportpath):
+                os.remove(webportpath)
 
         introfurl = os.path.join(introbase, 'introducer.furl')
